fastapi/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.0
astroid==3.3.4
bcrypt==4.2.0
black==24.8.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
dill==0.3.8
ecdsa==0.19.0
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
isort==5.13.2
Mako==1.3.5
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
peewee==3.17.6
platformdirs==4.3.6
py==1.11.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pylint==3.3.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tomlkit==0.13.2
typing_extensions==4.12.2
uvicorn==0.31.0
